requirements.txt
SpeechRecognition==3.10.1
gTTS==2.5.4
pyttsx3==2.90
pygame==2.6.1
wikipedia==1.4.0
pyjokes==0.8.3
requests==2.32.3
google-genai
sounddevice
